Added a prompt to ask you if you do indeed want to delete the application event log.

+++ Windows/EventLog.vb
@@ -536,8 +536,12 @@
     End Sub
 
     Private Sub btnCleanLogFile_Click(sender As Object, e As EventArgs) Handles btnCleanLogFile.Click
-        IO.File.Delete(Functions.eventLogFunctions.strLogFile)
-        Functions.eventLogFunctions.writeToApplicationLogFile(String.Format("Log file cleaned by user {0}.", Environment.UserName), EventLogEntryType.Information)
+        If MsgBox("Are you sure you want to delete the application event log?", MsgBoxStyle.Question + MsgBoxStyle.YesNo, "Are you sure?") = MsgBoxResult.Yes Then
+            IO.File.Delete(Functions.eventLogFunctions.strLogFile)
+            Functions.eventLogFunctions.writeToApplicationLogFile(String.Format("Log file cleaned by user {0}.", Environment.UserName), EventLogEntryType.Information)
+        Else
+            MsgBox("The application event log has not been deleted.", MsgBoxStyle.Information, Me.Text)
+        End If
     End Sub
 
     Private Sub logFileWatcher_Changed(sender As Object, e As IO.FileSystemEventArgs) Handles logFileWatcher.Changed
